+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to the "vscode-postgres" extension will be documented in thi
 
 ## [Unreleased]
 
+## [1.0.0] - 2018-05-05
+### Fixed
+- Connection button, now properly asks for the connection instead of the database.
+
+### Added
+- Syntax highlighting
+- Connection aware code completion (keywords, functions, tables, and fields)
+- In-line error detection powered by EXPLAIN (one error per query in editor)
+- *__Basic__* function signature support (connection aware)
+
 ## [0.0.6] - 2018-04-30
 ### Fixed
 - Repository link